pyproject.toml
[tool.poetry]
name = "gb4-wiki-gen"
version = "0.1.0"
description = "Generate content for gundambreaker.miraheze.org"
authors = ["Frederik Schumacher <fre-sch@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
jinja2 = "^3.1.4"
requests = "^2.32.3"
click = "^8.1.7"
python-slugify = "^8.0.4"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
